chore: extracted StreetDetailsModalListener out of map_screen.dart
